requirements.txt
azure-common==1.1.27
azure-core==1.21.1
azure-servicebus==7.5.0
azure-storage-blob==12.9.0
more-itertools==8.12.0
numpy==1.22.2
pandas==1.4.0
pyodbc==4.0.32
python-dotenv==0.19.2
SQLAlchemy==1.4.29
gql[all]==3.2.0